backend/requirements.txt
Django==5.2.11
djangorestframework==3.15.2
django-filter==24.3
dj-database-url==2.2.0
django-cors-headers==4.3.1
whitenoise==6.8.2
gunicorn==22.0.0
psycopg2-binary==2.9.10
cloudinary==1.41.0
python-dotenv==1.0.1
firebase-admin==6.4.0
